Addressing frequent question to CPSC advising
We're getting a lot of DSCI minor students expecting to be admitted to, e.g., CPSC 304 (or 322, 340, or 416) without the proper prereqs. The answer is a simple "no", but we'd like to avoid the questions if we can. This is a proposed extra FAQ entry to address that. Would be great if something like this could show on the main page, however. E.g., "As usual, to take a course listed in the upper-level requirements, you must meet that course's prereqs."